+++ src/CHANGELOG.md
@@ -11,6 +11,7 @@ v.3.11
 * [bug] Uninstall process was not compatible with PHP 8.1, throwing errors. Fixed Uninstall and Settings classes.
 * [bug] Better PHP 8.2 compatibility (function lumiere_parse_spans() in Movies class returns now string only instead of ( null | string)
 * [bug] The language form for people taxonomy was buggy. Fixed.
+* [bug] Title's "also known as" is back. It was unavailable since a very long time.
 * [technical] Lumière! is not executed in feeds anymore (created lumiere_prohibited_areas() in movie class)
 * [technical] Support for PHP < 8.0 totally dropped, removed home-made str_contains() compatibility meant to support older PHP versions.
 * [technical] Permissions for cache directory changed. Created with chmod 775, as it should.